Refactored drawing functions
