+++ frontend/app/cells/image/ImageCanvas/OutputClient.js
@@ -5,7 +5,7 @@ import { CanvasContext } from '../../../contexts/CanvasContext';
 import useLabels from '../../../../lib/hooks/useLabels';
 import { getColor, getContrastingColor } from '../../../../lib/generateChartColor';
 import truncateValue from '../../../../lib/truncateValue';
-import { processMask, drawMarker } from '../../../../lib/canvas';
+import { drawRegions, drawBoxes, drawLines, drawMask, drawMarkers } from '../../../../lib/canvas';
 import { isTagHidden, makeTag } from '../../../../lib/tags';
 import styles from './ImageCanvas.module.scss';
 import classNames from 'classnames/bind';
@@ -19,7 +19,7 @@ const ImageCanvasOutputClient = ({ assetId, dgid, timestamp, imageSrc }) => {
     const [isLoaded, setIsLoaded] = useState(false);
 
     const {
-	    layers,
+        layers,
         score,
         hiddenLabels,
     } = useLabels({ assetId, timestamp, dgid });
@@ -56,134 +56,44 @@ const ImageCanvasOutputClient = ({ assetId, dgid, timestamp, imageSrc }) => {
     }, [dimensions, imageScale]);
 
 
-    const drawLabels = useCallback(() => {
+    const drawAllAnnotations = useCallback(() => {
         if (layers) {
+	    // Globals:
             const alpha = 200; // get from slider?
             const ctx = labelCanvas.current.getContext("2d");
             ctx.font = "1em serif";
             ctx.textBaseline = "bottom";
             ctx.clearRect(0, 0, imgDims.width, imgDims.height);
+
             // Display any masks first:
-	    for (let layer of layers) {
-            for (let annotation of layer?.data) {
-                if (annotation.mask) {
-                    processMask(ctx, annotation, imgDims, hiddenLabels, layer.name, score, alpha);
+            for (let layer of layers) {
+                for (let annotation of layer?.data) {
+                    if (annotation.mask) {
+                        drawMask(ctx, annotation, imgDims, hiddenLabels, layer.name, score, alpha);
+                    }
                 }
             }
-	    }
-        // Next, draw all other annotations:
-	    for (let layer of layers) {
-	        for (let annotation of layer?.data) {
-                if (annotation.score && annotation.score <= score) continue;
-                if (!!annotation.points) {
-                    const points = annotation.points;
-                    if (!isTagHidden(hiddenLabels, makeTag(layer.name, annotation.label))) {
-                        for (let r = 0; r < points.length; r++) {
-                            const region = points[r];
-                            ctx.fillStyle = getColor(
-                                annotation.label
-                            );
-                            ctx.strokeStyle = getContrastingColor(ctx.fillStyle);
-                            ctx.lineWidth = 1;
-                            ctx.beginPath();
-                            ctx.moveTo(
-                                region[0] * imageScale,
-                                region[1] * imageScale
-                            );
-                            for (let i = 2; i < region.length; i += 2) {
-                                ctx.lineTo(
-                                    region[i] * imageScale,
-                                    region[i + 1] * imageScale
-                                );
-                            }
-                            ctx.closePath();
-                            ctx.fill();
-                            ctx.stroke();
-                        }
-                    }
-                } else if (!!annotation.boxes) {
-                    const boxes = annotation.boxes;
-                    if (!isTagHidden(hiddenLabels, makeTag(layer.name, annotation.label))) {
-                        for (let r = 0; r < boxes.length; r++) {
-                            const [x1, y1, x2, y2] = boxes[r];
-                            ctx.strokeStyle = getColor(
-                                annotation.label
-                            );
-                            if (r === 0) {
-                                let text = null;
+            // Next, draw all other annotations:
+            for (let layer of layers) {
+                for (let annotation of layer?.data) {
+                    if (annotation.score && annotation.score <= score)
+			continue;
 
-                                if (typeof annotation.score === "number") {
-                                    text = `${annotation.label}: ${truncateValue(annotation.score)}`;
-                                } else {
-                                    text = `${annotation.label}`;
-                                }
-
-                                const fontMetrics = ctx.measureText(text);
-                                const startX = x1 * imageScale - 1;
-                                const startY = y1 * imageScale;
-                                const border = 5;
-                                const width = fontMetrics.width + border * 2;
-                                const height = fontMetrics.fontBoundingBoxAscent + fontMetrics.fontBoundingBoxDescent + border * 2;
-
-                                // Draw text background box
-                                ctx.fillStyle = ctx.strokeStyle;
-                                ctx.beginPath();
-                                ctx.moveTo(startX, startY);
-                                ctx.lineTo( startX, startY - height);
-                                ctx.lineTo( startX + width, startY - height);
-                                ctx.lineTo( startX + width, startY);
-                                ctx.closePath();
-                                ctx.fill();
-
-                                // Draw the label:
-                                ctx.fillStyle = getContrastingColor(ctx.strokeStyle);
-                                ctx.fillText(text, startX + border, startY - border);
-                            }
-                            // Draw the bounding box
-                            ctx.lineWidth = 3;
-                            ctx.beginPath();
-                            ctx.moveTo(x1 * imageScale, y1 * imageScale);
-                            ctx.lineTo((x1 + x2) * imageScale, y1 * imageScale);
-                            ctx.lineTo((x1 + x2) * imageScale, (y1 + y2) * imageScale);
-                            ctx.lineTo(x1 * imageScale, (y1 + y2) * imageScale);
-                            ctx.closePath();
-                            ctx.stroke();
-                        }
+                    if (!!annotation.points) {
+			drawRegions(ctx, annotation.points, layer.name, annotation.label, imageScale, hiddenLabels);
+                    } else if (!!annotation.boxes) {
+			drawBoxes(ctx, annotation.boxes, layer.name, annotation.label, annotation.score, imageScale, hiddenLabels);
+                    } else if (annotation.lines) {
+			drawLines(ctx, annotation.lines, layer.name, annotation.label, imageScale, hiddenLabels);
+                    } else if (annotation.markers) {
+			drawMarkers(ctx, annotation.markers, layer.name, annotation.label, imageScale, hiddenLabels);
+                    } else if (annotation.mask) {
+			// skip here; already drawn
+                    } else {
+			console.log(`unknown annotation type: ${annotation}`);
                     }
-                } else if (annotation.lines) {
-                    const lines = annotation.lines;
-                    if (!isTagHidden(hiddenLabels, makeTag(layer.name, annotation.label))) {
-                        for (let r = 0; r < lines.length; r++) {
-                            const [x1, y1, x2, y2] = lines[r];
-                            ctx.strokeStyle = getColor(
-                                annotation.label
-                            );
-                            ctx.lineWidth = 3;
-                            ctx.beginPath();
-                            ctx.moveTo(x1 * imageScale, y1 * imageScale);
-                            ctx.lineTo(x2 * imageScale, y2 * imageScale);
-                            ctx.stroke();
-                        }
-                    }
-                } else if (annotation.markers) {
-                    const markers = annotation.markers;
-                    let marker = null;
-                    if (!isTagHidden(hiddenLabels, makeTag(layer.name, annotation.label))) {
-                        for (let r = 0; r < markers.length; r++) {
-                            const marker = markers[r];
-                            marker.color = getColor(
-                                annotation.label
-                            );
-                            drawMarker(ctx, marker, marker.x * imageScale, marker.y * imageScale);
-                        }
-                    }
-                } else if (annotation.mask) {
-                    // skip here; already drawn
-                } else {
-                    console.log(`unknown annotation type: ${annotation}`);
-                }
-             }
-	    }
+		}
+            }
         }
     }, [imageScale, score, hiddenLabels, layers, imgDims]);
 
@@ -210,9 +120,9 @@ const ImageCanvasOutputClient = ({ assetId, dgid, timestamp, imageSrc }) => {
 
     useEffect(() => {
         if (isLoaded) {
-            drawLabels();
+            drawAllAnnotations();
         }
-    }, [isLoaded, drawLabels]);
+    }, [isLoaded, drawAllAnnotations]);
 
 
     return (
@@ -263,6 +173,3 @@ const ImageCanvasOutputClient = ({ assetId, dgid, timestamp, imageSrc }) => {
 }
 
 export default ImageCanvasOutputClient;
-
-
-// for dev, add this to <canvas> title={`asset: ${assetId} imscale: ${imageScale} cd: ${labelCanvas.current?.height} x ${labelCanvas.current?.width} id: ${img.current?.height} x ${img.current?.width}`}
